safari-600.7-branch/LayoutTests/fast/dom/shadow/content-element-user-agent-shadow-expected.txt

PASS parsedElementShouldBeUnknown.tagName is 'WEBKITSHADOWCONTENT'
PASS 0 < parsedElementShouldBeUnknown.constructor.toString().indexOf('HTMLUnknownElement') is true
PASS contentInuserAgentShadow.tagName is 'WEBKITSHADOWCONTENT'
PASS 0 < contentInuserAgentShadow.constructor.toString().indexOf('HTMLElement') is true
PASS successfullyParsed is true

TEST COMPLETE
